backend/requirements.txt
strands-agents
fastapi
uvicorn
pytest
httpx
sqlalchemy
psycopg2-binary
bcrypt
pgvector
python-multipart
pyjwt
sse-starlette
boto3
